backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
google-genai==0.4.0
pillow==10.2.0
numpy==1.26.4
aiofiles==23.2.1
pydantic==2.6.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
